@@ -8,29 +8,13 @@
     "build": "next build && next export -o ./dist"
   },
   "dependencies": {
-    "@sanity/block-content-to-react": "^2.0.6",
-    "@sanity/client": "^0.140.11",
-    "@sanity/image-url": "^0.140.12",
     "aos": "^2.3.4",
-    "date-fns": "^1.30.1",
-    "formik": "^1.5.2",
-    "next": "^8.1.0",
-    "next-offline": "^3.3.8",
+    "formik": "^1.5.8",
+    "next": "^9.0.1",
+    "next-offline": "^4.0.2",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-ga": "^2.5.3",
-    "react-helmet": "^5.2.0",
-    "react-plx": "^1.3.13"
-  },
-  "devDependencies": {
-    "babel-eslint": "^10.0.1",
-    "core-js": "^3.0.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-babel": "^5.3.0",
-    "eslint-plugin-react": "^7.12.2",
-    "next-server": "^8.1.0",
-    "prop-types": "^15.7.2",
-    "react-is": "^16.8.6",
-    "regenerator-runtime": "^0.13.2"
+    "react-ga": "^2.6.0",
+    "react-helmet": "^5.2.1"
   }
 }
