@@ -1,28 +1,28 @@
-import React from 'react'
-import { Formik } from 'formik'
-import Input from './Input'
+import React from "react"
+import { Formik } from "formik"
+import Input from "./Input"
 
 export default class NewsletterForm extends React.Component {
-  constructor (props) {
+  constructor(props) {
     super(props)
     this.state = { honeypot: null }
     this.pleaseDont = React.createRef()
   }
 
-  componentDidMount () {
-    this.setState({ honeypot: document.getElementsByClassName('please-dont') })
+  componentDidMount() {
+    this.setState({ honeypot: document.getElementsByClassName("please-dont") })
   }
 
-  render () {
+  render() {
     return (
-      <React.Fragment>
+      <>
         <Formik
-          initialValues={{ name: '', email: '', message: '' }}
+          initialValues={{ name: "", email: "", message: "" }}
           validate={values => {
             let errors = {}
             if (!values.email) {
               errors.email =
-                'Email address is required, how are we supposed to contact you?'
+                "Email address is required, how are we supposed to contact you?"
             } else if (
               !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i.test(values.email)
             ) {
@@ -34,21 +34,21 @@ export default class NewsletterForm extends React.Component {
           onSubmit={(values, { setSubmitting, resetForm }) => {
             if (!this.state.honeypot.checked) {
               fetch(this.props.url, {
-                method: 'POST',
-                mode: 'no-cors',
+                method: "POST",
+                mode: "no-cors",
                 body: JSON.stringify(values),
                 headers: {
-                  'Content-Type': 'application/json'
+                  "Content-Type": "application/json"
                 }
               }).then(() => {
                 setSubmitting(false)
                 resetForm()
                 alert(
-                  'Thanks, we will get in touch soon! (this is the best method to display information after filling a form)'
+                  "Thanks, we will get in touch soon! (this is the best method to display information after filling a form)"
                 )
               })
             } else {
-              alert('Bad bad bot!')
+              alert("Bad bad bot!")
             }
           }}
         >
@@ -61,55 +61,55 @@ export default class NewsletterForm extends React.Component {
             handleBlur,
             isSubmitting
           }) => (
-            <form onSubmit={handleSubmit} className='gform'>
-              <span className='input'>
+            <form onSubmit={handleSubmit} className="gform">
+              <span className="input">
                 <input
-                  type='text'
-                  name='name'
-                  className='input__field name'
+                  type="text"
+                  name="name"
+                  className="input__field name"
                   onChange={handleChange}
                   onBlur={handleBlur}
-                  value={values.name || ''}
+                  value={values.name || ""}
                 />
-                <label htmlFor='email'>Your name</label>
+                <label htmlFor="email">Your name</label>
               </span>
-              <span className='input'>
+              <span className="input">
                 <input
-                  type='email'
-                  name='email'
-                  className='input__field email'
+                  type="email"
+                  name="email"
+                  className="input__field email"
                   onChange={handleChange}
                   onBlur={handleBlur}
-                  value={values.email || ''}
+                  value={values.email || ""}
                 />
-                <label htmlFor='email'>Your email</label>
+                <label htmlFor="email">Your email</label>
               </span>
-              <span className='input'>
+              <span className="input">
                 <textarea
-                  name='message'
-                  className='input__field textarea'
-                  cols='30'
-                  rows='5'
+                  name="message"
+                  className="input__field textarea"
+                  cols="30"
+                  rows="5"
                   onChange={handleChange}
                   onBlur={handleBlur}
-                  value={values.message || ''}
+                  value={values.message || ""}
                 />
-                <label htmlFor='email'>'sup</label>
+                <label htmlFor="email">'sup</label>
               </span>
               <input
-                className='please-dont'
-                type='checkbox'
-                value='1'
-                tabIndex='-1'
-                autoComplete='nope'
+                className="please-dont"
+                type="checkbox"
+                value="1"
+                tabIndex="-1"
+                autoComplete="nope"
               />
-              <div className={errors.email ? 'error' : 'error hidden'}>
+              <div className={errors.email ? "error" : "error hidden"}>
                 {errors.email}
               </div>
               <button
-                type='submit'
+                type="submit"
                 disabled={!values.email || errors.email}
-                className='button'
+                className="button"
               >
                 Submit
               </button>
@@ -156,7 +156,7 @@ export default class NewsletterForm extends React.Component {
             overflow: auto;
           }
 
-          .input__field:not([value='']):not(.textarea) + label,
+          .input__field:not([value=""]):not(.textarea) + label,
           .textarea:not(:empty) + label,
           .input__field:focus + label {
             transform: translateY(-100%);
@@ -204,7 +204,7 @@ export default class NewsletterForm extends React.Component {
             transform: translate(-500000px);
           }
         `}</style>
-      </React.Fragment>
+      </>
     )
   }
 }
