@@ -1,153 +1,153 @@
-import React from 'react'
-import AOS from 'aos'
+import React from "react"
+import AOS from "aos"
 
 class CssDesignAwards extends React.Component {
-  componentDidMount () {
+  componentDidMount() {
     AOS.init()
   }
 
-  render () {
+  render() {
     return (
       <a
-        href='https://www.cssdesignawards.com/sites/basement-studio/33829/'
-        target='_blank'
-        rel='noopener noreferrer'
-        data-aos='fade-up'
-        data-aos-offset='-100'
-        data-aos-delay='850'
-        data-aos-duration='700'
-        data-aos-easing='ease-in-out'
+        href="https://www.cssdesignawards.com/sites/basement-studio/33829/"
+        target="_blank"
+        rel="noopener noreferrer"
+        data-aos="fade-up"
+        data-aos-offset="-150"
+        data-aos-delay="850"
+        data-aos-duration="700"
+        data-aos-easing="ease-in-out"
       >
-        <div className='cssda-badge'>
+        <div className="cssda-badge">
           <svg
-            version='1.1'
-            className='cssda-badge-logo'
-            xmlns='http://www.w3.org/2000/svg'
-            xmlnsXlink='http://www.w3.org/1999/xlink'
-            x='0px'
-            y='0px'
-            viewBox='0 0 90 90'
-            xmlSpace='preserve'
+            version="1.1"
+            className="cssda-badge-logo"
+            xmlns="http://www.w3.org/2000/svg"
+            xmlnsXlink="http://www.w3.org/1999/xlink"
+            x="0px"
+            y="0px"
+            viewBox="0 0 90 90"
+            xmlSpace="preserve"
           >
             <path
-              d='M32.6,43.6c-0.3,0-0.5,0.1-0.7,0.4c-0.2,0.2-0.3,0.6-0.3,1c0,0.9,0.3,1.3,1,1.3c0.2,0,0.4,0,0.6-0.1
+              d="M32.6,43.6c-0.3,0-0.5,0.1-0.7,0.4c-0.2,0.2-0.3,0.6-0.3,1c0,0.9,0.3,1.3,1,1.3c0.2,0,0.4,0,0.6-0.1
            c0.2-0.1,0.4-0.1,0.6-0.2v1.1c-0.4,0.2-0.8,0.3-1.3,0.3c-0.7,0-1.2-0.2-1.6-0.6c-0.4-0.4-0.6-1-0.6-1.8c0-0.5,0.1-0.9,0.3-1.3
            c0.2-0.4,0.4-0.6,0.8-0.8c0.3-0.2,0.7-0.3,1.2-0.3c0.5,0,1,0.1,1.4,0.3l-0.4,1c-0.2-0.1-0.3-0.1-0.5-0.2
-           C33,43.6,32.8,43.6,32.6,43.6z'
+           C33,43.6,32.8,43.6,32.6,43.6z"
             />
             <path
-              d='M40,45.8c0,0.3-0.1,0.5-0.2,0.8c-0.1,0.2-0.4,0.4-0.6,0.5c-0.3,0.1-0.6,0.2-1,0.2c-0.3,0-0.6,0-0.8-0.1s-0.4-0.1-0.7-0.2
+              d="M40,45.8c0,0.3-0.1,0.5-0.2,0.8c-0.1,0.2-0.4,0.4-0.6,0.5c-0.3,0.1-0.6,0.2-1,0.2c-0.3,0-0.6,0-0.8-0.1s-0.4-0.1-0.7-0.2
            v-1.1c0.2,0.1,0.5,0.2,0.8,0.3s0.5,0.1,0.7,0.1c0.2,0,0.3,0,0.4-0.1c0.1-0.1,0.1-0.1,0.1-0.3c0-0.1,0-0.1-0.1-0.2
            c0,0-0.1-0.1-0.2-0.1c-0.1,0-0.3-0.1-0.6-0.3c-0.3-0.1-0.5-0.3-0.7-0.4c-0.2-0.1-0.3-0.3-0.3-0.4c-0.1-0.2-0.1-0.4-0.1-0.6
            c0-0.4,0.2-0.8,0.5-1c0.3-0.2,0.7-0.4,1.3-0.4c0.5,0,1,0.1,1.5,0.3l-0.4,1c-0.4-0.2-0.8-0.3-1.1-0.3c-0.2,0-0.3,0-0.4,0.1
-           C38,43.7,38,43.8,38,43.8c0,0.1,0,0.2,0.1,0.2c0.1,0.1,0.4,0.2,0.8,0.4c0.4,0.2,0.7,0.4,0.8,0.6C39.9,45.3,40,45.5,40,45.8z'
+           C38,43.7,38,43.8,38,43.8c0,0.1,0,0.2,0.1,0.2c0.1,0.1,0.4,0.2,0.8,0.4c0.4,0.2,0.7,0.4,0.8,0.6C39.9,45.3,40,45.5,40,45.8z"
             />
             <path
-              d='M45.9,45.8c0,0.3-0.1,0.5-0.2,0.8c-0.1,0.2-0.4,0.4-0.6,0.5c-0.3,0.1-0.6,0.2-1,0.2c-0.3,0-0.6,0-0.8-0.1
+              d="M45.9,45.8c0,0.3-0.1,0.5-0.2,0.8c-0.1,0.2-0.4,0.4-0.6,0.5c-0.3,0.1-0.6,0.2-1,0.2c-0.3,0-0.6,0-0.8-0.1
            s-0.4-0.1-0.7-0.2v-1.1c0.2,0.1,0.5,0.2,0.8,0.3s0.5,0.1,0.7,0.1c0.2,0,0.3,0,0.4-0.1c0.1-0.1,0.1-0.1,0.1-0.3
            c0-0.1,0-0.1-0.1-0.2c0,0-0.1-0.1-0.2-0.1c-0.1,0-0.3-0.1-0.6-0.3c-0.3-0.1-0.5-0.3-0.7-0.4c-0.2-0.1-0.3-0.3-0.3-0.4
            c-0.1-0.2-0.1-0.4-0.1-0.6c0-0.4,0.2-0.8,0.5-1c0.3-0.2,0.7-0.4,1.3-0.4c0.5,0,1,0.1,1.5,0.3l-0.4,1c-0.4-0.2-0.8-0.3-1.1-0.3
            c-0.2,0-0.3,0-0.4,0.1c-0.1,0.1-0.1,0.1-0.1,0.2c0,0.1,0,0.2,0.1,0.2c0.1,0.1,0.4,0.2,0.8,0.4c0.4,0.2,0.7,0.4,0.8,0.6
-           C45.9,45.3,45.9,45.5,45.9,45.8z'
+           C45.9,45.3,45.9,45.5,45.9,45.8z"
             />
             <path
-              d='M52.8,44.8c0,0.8-0.2,1.4-0.6,1.8c-0.4,0.4-1,0.6-1.8,0.6h-1.5v-4.7h1.6c0.8,0,1.3,0.2,1.7,0.6S52.8,44.1,52.8,44.8z
-            M51.5,44.9c0-0.4-0.1-0.7-0.3-1c-0.2-0.2-0.4-0.3-0.8-0.3h-0.4v2.6h0.3c0.4,0,0.7-0.1,0.8-0.3C51.4,45.7,51.5,45.3,51.5,44.9z'
+              d="M52.8,44.8c0,0.8-0.2,1.4-0.6,1.8c-0.4,0.4-1,0.6-1.8,0.6h-1.5v-4.7h1.6c0.8,0,1.3,0.2,1.7,0.6S52.8,44.1,52.8,44.8z
+            M51.5,44.9c0-0.4-0.1-0.7-0.3-1c-0.2-0.2-0.4-0.3-0.8-0.3h-0.4v2.6h0.3c0.4,0,0.7-0.1,0.8-0.3C51.4,45.7,51.5,45.3,51.5,44.9z"
             />
             <path
-              d='M58.7,47.3l-0.2-0.9h-1.5l-0.2,0.9h-1.4l1.5-4.7h1.7l1.5,4.7H58.7z M58.2,45.4L58,44.6c0-0.2-0.1-0.4-0.2-0.7
-           c-0.1-0.3-0.1-0.5-0.1-0.6c0,0.1-0.1,0.3-0.1,0.5c-0.1,0.2-0.2,0.7-0.4,1.5H58.2z'
+              d="M58.7,47.3l-0.2-0.9h-1.5l-0.2,0.9h-1.4l1.5-4.7h1.7l1.5,4.7H58.7z M58.2,45.4L58,44.6c0-0.2-0.1-0.4-0.2-0.7
+           c-0.1-0.3-0.1-0.5-0.1-0.6c0,0.1-0.1,0.3-0.1,0.5c-0.1,0.2-0.2,0.7-0.4,1.5H58.2z"
             />
-            <polygon points='65.8,39.9 65.8,39.2 51.2,39.2 46.4,34.4 41.7,39.2 39,39.2 49,29.2 38.5,29.2 28.4,39.2 24.6,39.2 24.6,39.9 	' />
+            <polygon points="65.8,39.9 65.8,39.2 51.2,39.2 46.4,34.4 41.7,39.2 39,39.2 49,29.2 38.5,29.2 28.4,39.2 24.6,39.2 24.6,39.9 	" />
             <polygon
-              points='24.6,49.9 24.6,50.6 39.1,50.6 43.8,55.3 48.6,50.6 51.2,50.6 41.2,60.6 51.7,60.6 57,55.4 61.7,50.6 65.8,50.6
-         65.8,49.9'
+              points="24.6,49.9 24.6,50.6 39.1,50.6 43.8,55.3 48.6,50.6 51.2,50.6 41.2,60.6 51.7,60.6 57,55.4 61.7,50.6 65.8,50.6
+         65.8,49.9"
             />
           </svg>
-          <div className='cssda-badge-content'>
+          <div className="cssda-badge-content">
             <svg
-              version='1.1'
-              className='cssda-badge-title'
-              xmlns='http://www.w3.org/2000/svg'
-              xmlnsXlink='http://www.w3.org/1999/xlink'
-              x='0px'
-              y='0px'
-              viewBox='0 0 90 90'
-              xmlSpace='preserve'
+              version="1.1"
+              className="cssda-badge-title"
+              xmlns="http://www.w3.org/2000/svg"
+              xmlnsXlink="http://www.w3.org/1999/xlink"
+              x="0px"
+              y="0px"
+              viewBox="0 0 90 90"
+              xmlSpace="preserve"
             >
               <path
-                d='M15.4,42.5l0,1.3l-2.8,0.7c-0.1,0-0.3,0.1-0.6,0.1c-0.3,0.1-0.5,0.1-0.6,0.1c0.1,0,0.3,0.1,0.6,0.1c0.3,0.1,0.5,0.1,0.6,0.1
+                d="M15.4,42.5l0,1.3l-2.8,0.7c-0.1,0-0.3,0.1-0.6,0.1c-0.3,0.1-0.5,0.1-0.6,0.1c0.1,0,0.3,0.1,0.6,0.1c0.3,0.1,0.5,0.1,0.6,0.1
            l2.7,0.8l0,1.3L10,48.2l0-1.1l2.9-0.6c0.5-0.1,1-0.2,1.4-0.2c-0.1,0-0.3-0.1-0.6-0.1C13.4,46,13.2,46,13,45.9l-2.9-0.8l0-1l2.9-0.7
-           c0.1,0,0.3-0.1,0.6-0.1c0.3,0,0.5-0.1,0.7-0.1c-0.2,0-0.4-0.1-0.7-0.1c-0.3-0.1-0.5-0.1-0.7-0.2l-2.8-0.7l0-1.1L15.4,42.5z'
+           c0.1,0,0.3-0.1,0.6-0.1c0.3,0,0.5-0.1,0.7-0.1c-0.2,0-0.4-0.1-0.7-0.1c-0.3-0.1-0.5-0.1-0.7-0.2l-2.8-0.7l0-1.1L15.4,42.5z"
               />
-              <path d='M16.6,36.4l-0.7,2.9l-5.1-1.2l0.7-2.9l0.9,0.2l-0.5,1.9l1.1,0.3l0.4-1.7l0.9,0.2l-0.4,1.7l1.3,0.3l0.5-1.9L16.6,36.4z' />
+              <path d="M16.6,36.4l-0.7,2.9l-5.1-1.2l0.7-2.9l0.9,0.2l-0.5,1.9l1.1,0.3l0.4-1.7l0.9,0.2l-0.4,1.7l1.3,0.3l0.5-1.9L16.6,36.4z" />
               <path
-                d='M12.6,31.9l0.7-1.5c0.3-0.7,0.6-1.1,1-1.3c0.3-0.2,0.7-0.2,1.1,0c0.3,0.1,0.5,0.3,0.6,0.5c0.1,0.2,0.1,0.4,0.1,0.7l0,0
+                d="M12.6,31.9l0.7-1.5c0.3-0.7,0.6-1.1,1-1.3c0.3-0.2,0.7-0.2,1.1,0c0.3,0.1,0.5,0.3,0.6,0.5c0.1,0.2,0.1,0.4,0.1,0.7l0,0
            c0.2-0.3,0.4-0.4,0.7-0.5c0.2,0,0.5,0,0.8,0.1c0.4,0.2,0.7,0.5,0.8,0.9c0.1,0.4,0,0.9-0.2,1.4l-0.8,1.8L12.6,31.9z M15,31.8l0.3-0.6
            c0.1-0.3,0.2-0.5,0.2-0.7c0-0.2-0.1-0.3-0.3-0.4c-0.2-0.1-0.3-0.1-0.5,0c-0.1,0.1-0.3,0.3-0.4,0.6l-0.2,0.5L15,31.8z M15.8,32.2
-           l1.2,0.6l0.3-0.7c0.1-0.3,0.2-0.5,0.1-0.7c0-0.2-0.2-0.3-0.4-0.4c-0.4-0.2-0.7,0-1,0.6L15.8,32.2z'
+           l1.2,0.6l0.3-0.7c0.1-0.3,0.2-0.5,0.1-0.7c0-0.2-0.2-0.3-0.4-0.4c-0.4-0.2-0.7,0-1,0.6L15.8,32.2z"
               />
               <path
-                d='M20.9,25.4c0.4,0.3,0.6,0.6,0.6,1.1c0,0.4-0.2,0.9-0.5,1.4c-0.3,0.5-0.7,0.8-1.1,1L19,28.2c0.3-0.2,0.6-0.4,0.8-0.5
+                d="M20.9,25.4c0.4,0.3,0.6,0.6,0.6,1.1c0,0.4-0.2,0.9-0.5,1.4c-0.3,0.5-0.7,0.8-1.1,1L19,28.2c0.3-0.2,0.6-0.4,0.8-0.5
            c0.2-0.2,0.4-0.3,0.5-0.5c0.1-0.2,0.2-0.4,0.2-0.5s-0.1-0.3-0.2-0.4c-0.1-0.1-0.2-0.1-0.3-0.1s-0.2,0-0.3,0.1s-0.4,0.2-0.7,0.3
            c-0.3,0.2-0.6,0.3-0.8,0.3c-0.2,0-0.4,0.1-0.6,0c-0.2,0-0.4-0.1-0.6-0.3c-0.4-0.3-0.6-0.6-0.6-1s0.1-0.8,0.4-1.3
            c0.2-0.2,0.3-0.4,0.5-0.6c0.2-0.2,0.4-0.3,0.7-0.5l0.5,0.8c-0.3,0.2-0.5,0.3-0.6,0.4c-0.1,0.1-0.3,0.3-0.4,0.4
            c-0.1,0.2-0.2,0.3-0.2,0.5c0,0.1,0.1,0.3,0.2,0.4c0.1,0.1,0.2,0.1,0.2,0.1c0.1,0,0.2,0,0.3,0c0.1,0,0.4-0.2,0.7-0.4
-           c0.5-0.3,0.9-0.4,1.2-0.4C20.3,25.1,20.6,25.2,20.9,25.4z'
+           c0.5-0.3,0.9-0.4,1.2-0.4C20.3,25.1,20.6,25.2,20.9,25.4z"
               />
-              <path d='M23.6,24.6l-3.7-3.7l0.8-0.8l3.7,3.7L23.6,24.6z' />
-              <path d='M27.9,20.9L27,21.5L24.5,18l-1.1,0.8l-0.5-0.7l3.2-2.3l0.5,0.7l-1.1,0.8L27.9,20.9z' />
-              <path d='M33.6,17.7L30.9,19l-2.3-4.7l2.7-1.3l0.4,0.8L30,14.6l0.5,1l1.6-0.8l0.4,0.8l-1.6,0.8l0.6,1.2l1.7-0.8L33.6,17.7z' />
+              <path d="M23.6,24.6l-3.7-3.7l0.8-0.8l3.7,3.7L23.6,24.6z" />
+              <path d="M27.9,20.9L27,21.5L24.5,18l-1.1,0.8l-0.5-0.7l3.2-2.3l0.5,0.7l-1.1,0.8L27.9,20.9z" />
+              <path d="M33.6,17.7L30.9,19l-2.3-4.7l2.7-1.3l0.4,0.8L30,14.6l0.5,1l1.6-0.8l0.4,0.8l-1.6,0.8l0.6,1.2l1.7-0.8L33.6,17.7z" />
               <path
-                d='M44,12.8c0.1,0.9-0.1,1.5-0.4,2.1c-0.4,0.5-1,0.8-1.8,0.9c-0.8,0.1-1.4-0.1-1.9-0.5s-0.8-1.1-0.8-1.9
+                d="M44,12.8c0.1,0.9-0.1,1.5-0.4,2.1c-0.4,0.5-1,0.8-1.8,0.9c-0.8,0.1-1.4-0.1-1.9-0.5s-0.8-1.1-0.8-1.9
            c-0.1-0.9,0.1-1.6,0.4-2.1c0.4-0.5,1-0.8,1.8-0.9c0.8-0.1,1.4,0.1,1.9,0.5C43.7,11.3,44,12,44,12.8z M40.2,13.2
            c0.1,0.6,0.2,1,0.5,1.3c0.3,0.3,0.6,0.4,1,0.3c0.9-0.1,1.3-0.7,1.1-1.9c-0.1-1.2-0.6-1.7-1.5-1.6c-0.4,0-0.8,0.2-1,0.5
-           C40.3,12.2,40.2,12.6,40.2,13.2z'
+           C40.3,12.2,40.2,12.6,40.2,13.2z"
               />
-              <path d='M47.7,15.7l-1.1-0.1l0.6-5.2l3,0.3l-0.1,0.9l-1.9-0.2L48,12.7l1.8,0.2l-0.1,0.9l-1.8-0.2L47.7,15.7z' />
-              <path d='M57.5,18.4l-1-0.5l1.8-3.9l-1.3-0.6l0.4-0.8l3.6,1.6L60.5,15l-1.3-0.6L57.5,18.4z' />
-              <path d='M64,22.5l-0.9-0.7l1.3-1.8l-1.7-1.2l-1.3,1.8L60.5,20l3.1-4.2l0.9,0.7l-1.2,1.7l1.7,1.2l1.2-1.7l0.9,0.7L64,22.5z' />
-              <path d='M67.8,26.5l-2-2.2l3.9-3.5l2,2.2l-0.7,0.6l-1.3-1.4L68.9,23l1.2,1.3l-0.7,0.6l-1.2-1.3l-1,0.9l1.3,1.4L67.8,26.5z' />
+              <path d="M47.7,15.7l-1.1-0.1l0.6-5.2l3,0.3l-0.1,0.9l-1.9-0.2L48,12.7l1.8,0.2l-0.1,0.9l-1.8-0.2L47.7,15.7z" />
+              <path d="M57.5,18.4l-1-0.5l1.8-3.9l-1.3-0.6l0.4-0.8l3.6,1.6L60.5,15l-1.3-0.6L57.5,18.4z" />
+              <path d="M64,22.5l-0.9-0.7l1.3-1.8l-1.7-1.2l-1.3,1.8L60.5,20l3.1-4.2l0.9,0.7l-1.2,1.7l1.7,1.2l1.2-1.7l0.9,0.7L64,22.5z" />
+              <path d="M67.8,26.5l-2-2.2l3.9-3.5l2,2.2l-0.7,0.6l-1.3-1.4L68.9,23l1.2,1.3l-0.7,0.6l-1.2-1.3l-1,0.9l1.3,1.4L67.8,26.5z" />
               <path
-                d='M75.4,34.8c-0.8,0.3-1.5,0.4-2.1,0.1c-0.6-0.3-1.1-0.8-1.5-1.7l-0.6-1.4l4.8-2l0.6,1.5c0.3,0.8,0.4,1.5,0.1,2.1
-           C76.6,34,76.1,34.5,75.4,34.8z M74.9,33.7c1-0.4,1.4-1.1,1-2l-0.2-0.5l-3.1,1.3l0.2,0.4C73.1,33.9,73.9,34.2,74.9,33.7z'
+                d="M75.4,34.8c-0.8,0.3-1.5,0.4-2.1,0.1c-0.6-0.3-1.1-0.8-1.5-1.7l-0.6-1.4l4.8-2l0.6,1.5c0.3,0.8,0.4,1.5,0.1,2.1
+           C76.6,34,76.1,34.5,75.4,34.8z M74.9,33.7c1-0.4,1.4-1.1,1-2l-0.2-0.5l-3.1,1.3l0.2,0.4C73.1,33.9,73.9,34.2,74.9,33.7z"
               />
               <path
-                d='M74.1,41.2l1.2-0.6l-0.3-1.9l-1.3-0.2l-0.2-1.2l5.5,0.9l0.2,1.3l-4.9,2.7L74.1,41.2z M76.1,40.2c1.1-0.5,1.6-0.8,1.8-0.9
-           c0.1-0.1,0.2-0.1,0.3-0.2c-0.3,0-1.1-0.1-2.4-0.3L76.1,40.2z'
+                d="M74.1,41.2l1.2-0.6l-0.3-1.9l-1.3-0.2l-0.2-1.2l5.5,0.9l0.2,1.3l-4.9,2.7L74.1,41.2z M76.1,40.2c1.1-0.5,1.6-0.8,1.8-0.9
+           c0.1-0.1,0.2-0.1,0.3-0.2c-0.3,0-1.1-0.1-2.4-0.3L76.1,40.2z"
               />
-              <path d='M77.4,46l2.1,1.2l0,1.2l-3.1-1.8l-2-0.1l0-1.1l2,0.1l3.3-1.6l0,1.2L77.4,46z' />
+              <path d="M77.4,46l2.1,1.2l0,1.2l-3.1-1.8l-2-0.1l0-1.1l2,0.1l3.3-1.6l0,1.2L77.4,46z" />
             </svg>
             <svg
-              version='1.1'
-              className='cssda-badge-award'
-              xmlns='http://www.w3.org/2000/svg'
-              xmlnsXlink='http://www.w3.org/1999/xlink'
-              x='0px'
-              y='0px'
-              viewBox='0 0 90 90'
-              xmlSpace='preserve'
+              version="1.1"
+              className="cssda-badge-award"
+              xmlns="http://www.w3.org/2000/svg"
+              xmlnsXlink="http://www.w3.org/1999/xlink"
+              x="0px"
+              y="0px"
+              viewBox="0 0 90 90"
+              xmlSpace="preserve"
             >
               <path
-                d='M25.5,73.7l-1.2-1l0.8-4.9l0,0c-0.4,0.6-0.8,1.1-1,1.3l-1.7,2.1l-0.8-0.7l3.6-4.4l1.2,1l-0.8,4.8l0,0
-   c0.4-0.6,0.7-1,0.9-1.2l1.7-2.1l0.8,0.7L25.5,73.7z'
+                d="M25.5,73.7l-1.2-1l0.8-4.9l0,0c-0.4,0.6-0.8,1.1-1,1.3l-1.7,2.1l-0.8-0.7l3.6-4.4l1.2,1l-0.8,4.8l0,0
+   c0.4-0.6,0.7-1,0.9-1.2l1.7-2.1l0.8,0.7L25.5,73.7z"
               />
               <path
-                d='M34.7,75.3c-0.4,0.9-0.9,1.4-1.5,1.7c-0.6,0.3-1.3,0.2-2.1-0.1c-0.8-0.4-1.3-0.9-1.5-1.5s-0.1-1.4,0.2-2.3s0.9-1.4,1.5-1.7
+                d="M34.7,75.3c-0.4,0.9-0.9,1.4-1.5,1.7c-0.6,0.3-1.3,0.2-2.1-0.1c-0.8-0.4-1.3-0.9-1.5-1.5s-0.1-1.4,0.2-2.3s0.9-1.4,1.5-1.7
    c0.6-0.3,1.3-0.2,2.1,0.1s1.3,0.9,1.5,1.5C35.2,73.7,35.1,74.4,34.7,75.3z M30.9,73.6c-0.3,0.6-0.3,1.1-0.3,1.5s0.4,0.7,0.8,0.9
-   c0.9,0.4,1.6,0,2.1-1.2c0.5-1.2,0.3-1.9-0.5-2.3c-0.4-0.2-0.8-0.2-1.2,0S31.2,73,30.9,73.6z'
+   c0.9,0.4,1.6,0,2.1-1.2c0.5-1.2,0.3-1.9-0.5-2.3c-0.4-0.2-0.8-0.2-1.2,0S31.2,73,30.9,73.6z"
               />
               <path
-                d='M39.5,79.3l-0.7-4.6l0,0c-0.1,0.9-0.1,1.5-0.2,1.8l-0.4,2.6l-1.1-0.1l0.8-5.6l1.6,0.2l0.7,4.5l0,0l2-4.1l1.6,0.2l-0.8,5.6
-   L42,79.6l0.4-2.7c0-0.1,0-0.3,0.1-0.4s0.1-0.6,0.2-1.3l0,0l-2.1,4.2L39.5,79.3z'
+                d="M39.5,79.3l-0.7-4.6l0,0c-0.1,0.9-0.1,1.5-0.2,1.8l-0.4,2.6l-1.1-0.1l0.8-5.6l1.6,0.2l0.7,4.5l0,0l2-4.1l1.6,0.2l-0.8,5.6
+   L42,79.6l0.4-2.7c0-0.1,0-0.3,0.1-0.4s0.1-0.6,0.2-1.3l0,0l-2.1,4.2L39.5,79.3z"
               />
-              <path d='M47,79.6L46.5,74l1.2-0.1l0.4,5.6L47,79.6z' />
+              <path d="M47,79.6L46.5,74l1.2-0.1l0.4,5.6L47,79.6z" />
               <path
-                d='M56.7,77.7l-1.5,0.4l-3.6-3.4l0,0c0.3,0.7,0.4,1.2,0.5,1.5l0.7,2.6l-1,0.3l-1.6-5.4l1.5-0.4l3.5,3.4l0,0
-   c-0.2-0.7-0.4-1.2-0.5-1.5l-0.7-2.6l1-0.3L56.7,77.7z'
+                d="M56.7,77.7l-1.5,0.4l-3.6-3.4l0,0c0.3,0.7,0.4,1.2,0.5,1.5l0.7,2.6l-1,0.3l-1.6-5.4l1.5-0.4l3.5,3.4l0,0
+   c-0.2-0.7-0.4-1.2-0.5-1.5l-0.7-2.6l1-0.3L56.7,77.7z"
               />
-              <path d='M63,74.5l-2.8,1.6l-2.8-4.9l2.8-1.6l0.5,0.9l-1.8,1l0.6,1.1l1.7-0.9l0.5,0.9L60,73.3l0.7,1.3l1.8-1L63,74.5z' />
-              <path d='M68.3,70.3l-2.5,2.1l-3.7-4.3l2.5-2.1l0.6,0.7l-1.6,1.3l0.8,0.9l1.4-1.3l0.6,0.7l-1.4,1.3l1,1.1l1.6-1.3L68.3,70.3z' />
+              <path d="M63,74.5l-2.8,1.6l-2.8-4.9l2.8-1.6l0.5,0.9l-1.8,1l0.6,1.1l1.7-0.9l0.5,0.9L60,73.3l0.7,1.3l1.8-1L63,74.5z" />
+              <path d="M68.3,70.3l-2.5,2.1l-3.7-4.3l2.5-2.1l0.6,0.7l-1.6,1.3l0.8,0.9l1.4-1.3l0.6,0.7l-1.4,1.3l1,1.1l1.6-1.3L68.3,70.3z" />
             </svg>
           </div>
         </div>
