@@ -1,21 +1,15 @@
-import React from 'react'
+import React from "react"
 
-class Logo extends React.Component {
-  constructor (props) {
-    super(props)
-  }
-
-  render () {
-    return (
-      <svg viewBox='0 0 200 33' xmlns='http://www.w3.org/2000/svg'>
-        <path
-          d='M16.885 21.409c0-3.722-1.95-6.111-5.06-6.111-3.527 0-5.43 2.94-5.43 5.467v1.425c0 2.71 1.996 5.513 5.384 5.513 3.017 0 5.106-2.48 5.106-6.294m6.451-.276c0 6.984-3.945 11.854-10.118 11.854-3.434 0-5.616-1.424-6.87-3.537l-.138 3.308H.037V.135h6.359v13.279c1.345-1.976 3.666-3.4 7.054-3.4 5.755 0 9.886 3.997 9.886 11.119zm21.814-1.7v13.325h-5.894l-.14-3.079c-.881 1.425-3.016 3.308-7.425 3.308-4.92 0-7.565-2.435-7.565-6.202 0-5.468 5.662-7.077 11.835-7.077h3.017v-.78c0-2.528-1.44-4.136-4.27-4.136-2.414 0-3.992 1.149-4.224 3.125H24.73c.418-5.009 4.363-7.904 10.164-7.904 6.823 0 10.257 3.63 10.257 9.42zm11.511 13.554c-6.405 0-9.654-2.573-9.7-7.26l5.43.047c.14 1.838 1.393 2.848 4.177 2.848 2.692 0 3.852-.78 3.806-2.067 0-4.227-12.949-.781-12.949-9.88 0-4.134 3.713-6.662 9.7-6.662 5.802 0 9.05 2.665 9.097 7.122H60.7c-.14-1.837-1.392-2.756-3.62-2.756-2.414 0-3.48.827-3.48 2.113 0 4.135 12.948.506 12.948 9.879.046 4.181-3.806 6.616-9.886 6.616zm33-10.108H73.649c.464 3.446 2.506 5.192 5.476 5.192 3.063 0 4.177-1.562 4.549-2.757h5.894c-.743 3.906-3.899 7.673-10.768 7.673-6.915 0-11.278-4.732-11.278-11.44 0-6.8 4.78-11.534 11.371-11.534 6.915 0 10.768 4.78 10.768 11.671v1.195zm37.409-4.32v14.199h-6.405v-12.82c0-2.849-1.44-4.457-3.76-4.457-2.552 0-4.316 1.884-4.316 4.687v12.59h-6.359v-12.82c0-2.803-1.392-4.457-3.76-4.457-2.552 0-4.315 1.884-4.315 4.687v12.59h-6.359V10.243h6.127l.185 3.263c1.3-2.068 3.528-3.493 6.684-3.493 3.666 0 6.08 1.884 7.147 4.687 1.16-2.71 3.62-4.687 7.38-4.687 5.059 0 7.75 3.538 7.75 8.547zm24.042 4.32H135.1c.464 3.446 2.506 5.192 5.477 5.192 3.063 0 4.177-1.562 4.548-2.757h5.895c-.743 3.906-3.9 7.673-10.768 7.673-6.916 0-11.279-4.732-11.279-11.44 0-6.8 4.781-11.534 11.372-11.534 6.915 0 10.767 4.78 10.767 11.671v1.195zm23.485-4.227v14.106h-6.358V20.076c0-2.849-1.346-4.595-4.038-4.595-2.785 0-4.595 1.884-4.595 4.687v12.59h-6.358V10.243h6.126l.186 3.309c1.346-2.114 3.62-3.539 7.101-3.539 5.152 0 7.936 3.584 7.936 8.639zm10.815-3.4v8.224c0 2.665.975 3.998 4.316 4.044h2.494v5.238h-3.468c-6.359-.046-9.7-3.308-9.7-9.052v-8.454h-3.342v-5.009h3.713V3.856h5.987v6.387h6.776v5.009h-6.776zM38.978 24.579v-1.424h-3.296c-3.527 0-5.384.919-5.384 2.94 0 1.425 1.115 2.344 3.342 2.344 3.342 0 5.338-2.068 5.338-3.86zm34.763-5.33h9.979c-.418-3.216-2.181-4.549-4.78-4.549-2.553 0-4.595 1.424-5.199 4.549zm61.452 0h9.978c-.417-3.216-2.18-4.549-4.78-4.549-2.553 0-4.595 1.424-5.198 4.549zm59.53 8.251H200v5.225h-5.278V27.5z'
-          fill='#FFF'
-          fillRule='evenodd'
-        />
-      </svg>
-    )
-  }
+function Logo({ color = "#fff" }) {
+  return (
+    <svg viewBox="0 0 200 33" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M16.885 21.409c0-3.722-1.95-6.111-5.06-6.111-3.527 0-5.43 2.94-5.43 5.467v1.425c0 2.71 1.996 5.513 5.384 5.513 3.017 0 5.106-2.48 5.106-6.294m6.451-.276c0 6.984-3.945 11.854-10.118 11.854-3.434 0-5.616-1.424-6.87-3.537l-.138 3.308H.037V.135h6.359v13.279c1.345-1.976 3.666-3.4 7.054-3.4 5.755 0 9.886 3.997 9.886 11.119zm21.814-1.7v13.325h-5.894l-.14-3.079c-.881 1.425-3.016 3.308-7.425 3.308-4.92 0-7.565-2.435-7.565-6.202 0-5.468 5.662-7.077 11.835-7.077h3.017v-.78c0-2.528-1.44-4.136-4.27-4.136-2.414 0-3.992 1.149-4.224 3.125H24.73c.418-5.009 4.363-7.904 10.164-7.904 6.823 0 10.257 3.63 10.257 9.42zm11.511 13.554c-6.405 0-9.654-2.573-9.7-7.26l5.43.047c.14 1.838 1.393 2.848 4.177 2.848 2.692 0 3.852-.78 3.806-2.067 0-4.227-12.949-.781-12.949-9.88 0-4.134 3.713-6.662 9.7-6.662 5.802 0 9.05 2.665 9.097 7.122H60.7c-.14-1.837-1.392-2.756-3.62-2.756-2.414 0-3.48.827-3.48 2.113 0 4.135 12.948.506 12.948 9.879.046 4.181-3.806 6.616-9.886 6.616zm33-10.108H73.649c.464 3.446 2.506 5.192 5.476 5.192 3.063 0 4.177-1.562 4.549-2.757h5.894c-.743 3.906-3.899 7.673-10.768 7.673-6.915 0-11.278-4.732-11.278-11.44 0-6.8 4.78-11.534 11.371-11.534 6.915 0 10.768 4.78 10.768 11.671v1.195zm37.409-4.32v14.199h-6.405v-12.82c0-2.849-1.44-4.457-3.76-4.457-2.552 0-4.316 1.884-4.316 4.687v12.59h-6.359v-12.82c0-2.803-1.392-4.457-3.76-4.457-2.552 0-4.315 1.884-4.315 4.687v12.59h-6.359V10.243h6.127l.185 3.263c1.3-2.068 3.528-3.493 6.684-3.493 3.666 0 6.08 1.884 7.147 4.687 1.16-2.71 3.62-4.687 7.38-4.687 5.059 0 7.75 3.538 7.75 8.547zm24.042 4.32H135.1c.464 3.446 2.506 5.192 5.477 5.192 3.063 0 4.177-1.562 4.548-2.757h5.895c-.743 3.906-3.9 7.673-10.768 7.673-6.916 0-11.279-4.732-11.279-11.44 0-6.8 4.781-11.534 11.372-11.534 6.915 0 10.767 4.78 10.767 11.671v1.195zm23.485-4.227v14.106h-6.358V20.076c0-2.849-1.346-4.595-4.038-4.595-2.785 0-4.595 1.884-4.595 4.687v12.59h-6.358V10.243h6.126l.186 3.309c1.346-2.114 3.62-3.539 7.101-3.539 5.152 0 7.936 3.584 7.936 8.639zm10.815-3.4v8.224c0 2.665.975 3.998 4.316 4.044h2.494v5.238h-3.468c-6.359-.046-9.7-3.308-9.7-9.052v-8.454h-3.342v-5.009h3.713V3.856h5.987v6.387h6.776v5.009h-6.776zM38.978 24.579v-1.424h-3.296c-3.527 0-5.384.919-5.384 2.94 0 1.425 1.115 2.344 3.342 2.344 3.342 0 5.338-2.068 5.338-3.86zm34.763-5.33h9.979c-.418-3.216-2.181-4.549-4.78-4.549-2.553 0-4.595 1.424-5.199 4.549zm61.452 0h9.978c-.417-3.216-2.18-4.549-4.78-4.549-2.553 0-4.595 1.424-5.198 4.549zm59.53 8.251H200v5.225h-5.278V27.5z"
+        fill={color}
+        fillRule="evenodd"
+      />
+    </svg>
+  )
 }
 
 export default Logo
