@@ -1,138 +1,133 @@
-import React from 'react'
+import React from "react"
 
-class LogoMobile extends React.Component {
-  constructor (props) {
-    super(props)
-  }
-  render () {
-    return (
-      <svg
-        viewBox='0 0 300 60'
-        xmlns='http://www.w3.org/2000/svg'
-        xmlnsXlink='http://www.w3.org/1999/xlink'
-      >
-        <defs>
-          <path
-            d='M24.45 34.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V3h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V31.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V31.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V17.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V32.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V17.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V25.078h-4.849v-7.314h5.388V8.436h8.69v9.328h9.834v7.314h-9.835zM56.513 38.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V50.6h-7.66v-7.632z'
-            id='e'
-          />
-          <pattern
-            id='f'
-            width='3.24'
-            height='3.24'
-            x='-3.24'
-            y='-.24'
-            patternUnits='userSpaceOnUse'
-          >
-            <use xlinkHref='#a' transform='scale(.40502)' />
-          </pattern>
-          <image
-            id='a'
-            width='8'
-            height='8'
-            xlinkHref='data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAICAYAAADED76LAAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAmSURBVBgZY2BgYPgPxLgAPjlcegiIg4xEMZaJgAYqSaPYiWbmfwAvDQX9S7RrNQAAAABJRU5ErkJggg=='
-          />
-          <linearGradient x1='50%' y1='0%' x2='50%' y2='100%' id='g'>
-            <stop stopColor='#FFF' stopOpacity='.5' offset='0%' />
-            <stop stopOpacity='.5' offset='100%' />
-          </linearGradient>
-          <path
-            d='M26.464 31.07c0-5.435-2.83-8.924-7.342-8.924-5.12 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.812 8.053 4.379 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.725 17.314-14.684 17.314-4.984 0-8.15-2.08-9.969-5.167l-.202 4.831H2.013V0h9.228v19.394c1.953-2.885 5.321-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.659-2.483v19.461H58.93l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.979-3.557-10.979-9.06 0-7.985 8.217-10.334 17.176-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.502 0-5.792 1.677-6.129 4.563h-8.352c.606-7.315 6.331-11.543 14.751-11.543 9.9 0 14.886 5.302 14.886 13.757zM84.19 47.982c-9.295 0-14.01-3.758-14.078-10.603l7.881.067c.202 2.684 2.021 4.16 6.062 4.16 3.907 0 5.59-1.14 5.523-3.02 0-6.173-18.792-1.14-18.792-14.427 0-6.04 5.389-9.73 14.078-9.73 8.419 0 13.134 3.891 13.201 10.4H90.05c-.202-2.683-2.021-4.026-5.254-4.026-3.503 0-5.051 1.208-5.051 3.087 0 6.04 18.792.739 18.792 14.428.067 6.107-5.523 9.664-14.347 9.664zm47.891-14.764h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.554c-1.078 5.704-5.658 11.207-15.626 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.088-6.51-5.456-6.51-3.704 0-6.264 2.751-6.264 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.704 0-6.264 2.751-6.264 6.845v18.387h-9.227V14.764h8.89l.27 4.764c1.886-3.02 5.119-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.684-3.96 5.254-6.845 10.71-6.845 7.342 0 11.248 5.167 11.248 12.482zm34.893 6.308h-23.238c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.083-6.174v20.602h-9.228V29.125c0-4.161-1.953-6.711-5.86-6.711-4.041 0-6.668 2.751-6.668 6.845v18.387h-9.228V14.764h8.89l.27 4.832c1.954-3.087 5.254-5.168 10.306-5.168 7.477 0 11.518 5.235 11.518 12.616zm15.695-4.966v12.013c0 3.892 1.414 5.838 6.264 5.905h3.62v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.85v-7.314h5.389V5.436h8.689v9.328h9.834v7.314h-9.834zM58.526 35.701v-2.08h-4.782c-5.12 0-7.814 1.342-7.814 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.452-7.784h14.481c-.606-4.698-3.165-6.644-6.938-6.644-3.704 0-6.668 2.08-7.543 6.644zm89.182 0h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z'
-            id='h'
-          />
-          <pattern
-            id='i'
-            width='3.685'
-            height='3.685'
-            x='-1.671'
-            y='-3.685'
-            patternUnits='userSpaceOnUse'
-          >
-            <use xlinkHref='#b' transform='scale(.61412)' />
-          </pattern>
-          <image
-            id='b'
-            width='6'
-            height='6'
-            xlinkHref='data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAYAAAAGCAYAAADgzO9IAAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAWSURBVAgdY2BgYPgPxUAKAZgQTPqzACB/AgGN23nBAAAAAElFTkSuQmCC'
-          />
-          <path
-            d='M24.45 31.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V0h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V14.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V14.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.849v-7.314h5.388V5.436h8.69v9.328h9.834v7.314h-9.835zM56.513 35.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z'
-            id='j'
-          />
-          <pattern
-            id='k'
-            width='3.5'
-            height='3.5'
-            x='-3.5'
-            y='-3.5'
-            patternUnits='userSpaceOnUse'
-          >
-            <use xlinkHref='#c' transform='scale(.5)' />
-          </pattern>
-          <image
-            id='c'
-            width='7'
-            height='7'
-            xlinkHref='data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAcAAAAHCAYAAADEUlfTAAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAeSURBVAgdY2TADv5jF2ZgAElglaSHBMhRWC0HiQMAxgkM9QW5fScAAAAASUVORK5CYII='
-          />
-          <path
-            d='M24.45 31.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V0h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V14.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V14.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.849v-7.314h5.388V5.436h8.69v9.328h9.834v7.314h-9.835zM56.513 35.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z'
-            id='l'
-          />
-          <pattern
-            id='m'
-            width='1.2'
-            height='1.2'
-            x='-1.2'
-            y='-1.2'
-            patternUnits='userSpaceOnUse'
-          >
-            <use xlinkHref='#d' transform='scale(.3)' />
-          </pattern>
-          <image
-            id='d'
-            width='4'
-            height='4'
-            xlinkHref='data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAQAAAAECAYAAACp8Z5+AAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAXSURBVAgdY2BgYPgPxUCKgYEJTJJEAAB8LgIB7J8/KwAAAABJRU5ErkJggg=='
-          />
-          <linearGradient x1='50%' y1='0%' x2='50%' y2='100%' id='n'>
-            <stop stopOpacity='.784' offset='0%' />
-            <stop stopOpacity='.931' offset='100%' />
-          </linearGradient>
-        </defs>
-        <g fill='none' fillRule='evenodd'>
-          <g transform='translate(.006 8.845)'>
-            <g opacity='.5'>
-              <use fill='#FFF' xlinkHref='#e' />
-              <use fill='url(#f)' xlinkHref='#e' />
-            </g>
-            <use fill='url(#g)' xlinkHref='#h' />
-            <use fill='url(#i)' xlinkHref='#h' />
+function LogoGlitched() {
+  return (
+    <svg
+      viewBox="0 0 300 60"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+    >
+      <defs>
+        <path
+          d="M24.45 34.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V3h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V31.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V31.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V17.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V32.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V17.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V25.078h-4.849v-7.314h5.388V8.436h8.69v9.328h9.834v7.314h-9.835zM56.513 38.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V50.6h-7.66v-7.632z"
+          id="e"
+        />
+        <pattern
+          id="f"
+          width="3.24"
+          height="3.24"
+          x="-3.24"
+          y="-.24"
+          patternUnits="userSpaceOnUse"
+        >
+          <use xlinkHref="#a" transform="scale(.40502)" />
+        </pattern>
+        <image
+          id="a"
+          width="8"
+          height="8"
+          xlinkHref="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAgAAAAICAYAAADED76LAAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAmSURBVBgZY2BgYPgPxLgAPjlcegiIg4xEMZaJgAYqSaPYiWbmfwAvDQX9S7RrNQAAAABJRU5ErkJggg=="
+        />
+        <linearGradient x1="50%" y1="0%" x2="50%" y2="100%" id="g">
+          <stop stopColor="#FFF" stopOpacity=".5" offset="0%" />
+          <stop stopOpacity=".5" offset="100%" />
+        </linearGradient>
+        <path
+          d="M26.464 31.07c0-5.435-2.83-8.924-7.342-8.924-5.12 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.812 8.053 4.379 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.725 17.314-14.684 17.314-4.984 0-8.15-2.08-9.969-5.167l-.202 4.831H2.013V0h9.228v19.394c1.953-2.885 5.321-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.659-2.483v19.461H58.93l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.979-3.557-10.979-9.06 0-7.985 8.217-10.334 17.176-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.502 0-5.792 1.677-6.129 4.563h-8.352c.606-7.315 6.331-11.543 14.751-11.543 9.9 0 14.886 5.302 14.886 13.757zM84.19 47.982c-9.295 0-14.01-3.758-14.078-10.603l7.881.067c.202 2.684 2.021 4.16 6.062 4.16 3.907 0 5.59-1.14 5.523-3.02 0-6.173-18.792-1.14-18.792-14.427 0-6.04 5.389-9.73 14.078-9.73 8.419 0 13.134 3.891 13.201 10.4H90.05c-.202-2.683-2.021-4.026-5.254-4.026-3.503 0-5.051 1.208-5.051 3.087 0 6.04 18.792.739 18.792 14.428.067 6.107-5.523 9.664-14.347 9.664zm47.891-14.764h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.554c-1.078 5.704-5.658 11.207-15.626 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.088-6.51-5.456-6.51-3.704 0-6.264 2.751-6.264 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.704 0-6.264 2.751-6.264 6.845v18.387h-9.227V14.764h8.89l.27 4.764c1.886-3.02 5.119-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.684-3.96 5.254-6.845 10.71-6.845 7.342 0 11.248 5.167 11.248 12.482zm34.893 6.308h-23.238c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.083-6.174v20.602h-9.228V29.125c0-4.161-1.953-6.711-5.86-6.711-4.041 0-6.668 2.751-6.668 6.845v18.387h-9.228V14.764h8.89l.27 4.832c1.954-3.087 5.254-5.168 10.306-5.168 7.477 0 11.518 5.235 11.518 12.616zm15.695-4.966v12.013c0 3.892 1.414 5.838 6.264 5.905h3.62v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.85v-7.314h5.389V5.436h8.689v9.328h9.834v7.314h-9.834zM58.526 35.701v-2.08h-4.782c-5.12 0-7.814 1.342-7.814 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.452-7.784h14.481c-.606-4.698-3.165-6.644-6.938-6.644-3.704 0-6.668 2.08-7.543 6.644zm89.182 0h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z"
+          id="h"
+        />
+        <pattern
+          id="i"
+          width="3.685"
+          height="3.685"
+          x="-1.671"
+          y="-3.685"
+          patternUnits="userSpaceOnUse"
+        >
+          <use xlinkHref="#b" transform="scale(.61412)" />
+        </pattern>
+        <image
+          id="b"
+          width="6"
+          height="6"
+          xlinkHref="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAYAAAAGCAYAAADgzO9IAAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAWSURBVAgdY2BgYPgPxUAKAZgQTPqzACB/AgGN23nBAAAAAElFTkSuQmCC"
+        />
+        <path
+          d="M24.45 31.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V0h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V14.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V14.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.849v-7.314h5.388V5.436h8.69v9.328h9.834v7.314h-9.835zM56.513 35.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z"
+          id="j"
+        />
+        <pattern
+          id="k"
+          width="3.5"
+          height="3.5"
+          x="-3.5"
+          y="-3.5"
+          patternUnits="userSpaceOnUse"
+        >
+          <use xlinkHref="#c" transform="scale(.5)" />
+        </pattern>
+        <image
+          id="c"
+          width="7"
+          height="7"
+          xlinkHref="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAcAAAAHCAYAAADEUlfTAAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAeSURBVAgdY2TADv5jF2ZgAElglaSHBMhRWC0HiQMAxgkM9QW5fScAAAAASUVORK5CYII="
+        />
+        <path
+          d="M24.45 31.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V0h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V14.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V14.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.849v-7.314h5.388V5.436h8.69v9.328h9.834v7.314h-9.835zM56.513 35.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z"
+          id="l"
+        />
+        <pattern
+          id="m"
+          width="1.2"
+          height="1.2"
+          x="-1.2"
+          y="-1.2"
+          patternUnits="userSpaceOnUse"
+        >
+          <use xlinkHref="#d" transform="scale(.3)" />
+        </pattern>
+        <image
+          id="d"
+          width="4"
+          height="4"
+          xlinkHref="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAQAAAAECAYAAACp8Z5+AAAABGdBTUEAALGPC/xhBQAAAYRpQ0NQSUNDIFByb2ZpbGUAACiRlZE/SMNAFMa/VKRL/TMUEV0CSqcqtYJUEKHtIEKHWjqo6JAm6R+oSUjOqugo6FhwEF0Uuzg461pw1UlRBHFwci+6aInvkmorouIH4f3y7t29e98BnmPJMIqebmBJY2ZqKibOzs2L3kd40Q9HkmwZ0WQywfkjftXLDQQer4b4Wd/Xf5VPUS0ZEPzETDZMRrxNPLDCDM4nxH6TLkVc5Zxz+ZpzxuUnpyadihO/EXeW5Bzt9fiIQ5pS0IgniCflvKQQbxAHMy31uRZ27+OoLy4VCxlTYqoicmvielE3LUOS1X8O+ZeYusp4jOvGmlnI5ZkYJSdVcVqTh4NiODQSAfi7uNW1GcdvoeeimdMPgcgz0FZu5jK7wNkW0HvXzA0eAF2bwOm5vGyWGu0FUHfXs8a/8En4iV1fHY0BlVsgvQ4kLoG9fSCQpT4LQLKD8uM04WLzc2VlR8PuWb4Y0P5g27UA4N0B6mXbfj2y7XqF5rkHqto7PMVvnmqUz74AAAAXSURBVAgdY2BgYPgPxUCKgYEJTJJEAAB8LgIB7J8/KwAAAABJRU5ErkJggg=="
+        />
+        <linearGradient x1="50%" y1="0%" x2="50%" y2="100%" id="n">
+          <stop stopOpacity=".784" offset="0%" />
+          <stop stopOpacity=".931" offset="100%" />
+        </linearGradient>
+      </defs>
+      <g fill="none" fillRule="evenodd">
+        <g transform="translate(.006 8.845)">
+          <g opacity=".5">
+            <use fill="#FFF" xlinkHref="#e" />
+            <use fill="url(#f)" xlinkHref="#e" />
           </g>
-          <g transform='translate(4.037 7.058)'>
-            <use fill='url(#g)' xlinkHref='#j' />
-            <use fill='url(#k)' xlinkHref='#j' />
-          </g>
-          <g transform='translate(7.062 5.066)'>
-            <use fill='url(#g)' xlinkHref='#l' />
-            <use fill='url(#m)' xlinkHref='#l' />
-          </g>
-          <path
-            d='M24.45 31.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V0h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V14.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V14.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.849v-7.314h5.388V5.436h8.69v9.328h9.834v7.314h-9.835zM56.513 35.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z'
-            transform='translate(5.422 4.311)'
-            fill='url(#n)'
-          />
-          <path
-            d='M32.958 31.382c0-5.435-2.829-8.925-7.342-8.925-5.119 0-7.88 4.295-7.88 7.986v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.624 7.41-9.194m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.969-5.167l-.202 4.832H8.508V.311h9.228v19.394c1.953-2.885 5.32-4.965 10.238-4.965 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.482v19.46h-8.554l-.202-4.495c-1.28 2.08-4.378 4.831-10.777 4.831-7.14 0-10.98-3.556-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.691-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.314 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.758zm16.706 19.796c-9.296 0-14.01-3.758-14.078-10.603l7.88.068c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.792-1.14-18.792-14.428 0-6.04 5.388-9.73 14.077-9.73 8.42 0 13.135 3.892 13.202 10.401h-8.015c-.202-2.684-2.021-4.026-5.254-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.738 18.793 14.428.067 6.107-5.524 9.663-14.347 9.663zm47.89-14.763h-23.237c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.282 6.6-4.027h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.931 6.938-16.843 16.503-16.843 10.036 0 15.627 6.979 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V29.235c0-4.161-2.087-6.51-5.455-6.51-3.705 0-6.264 2.752-6.264 6.845v18.388h-9.228V29.235c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.752-6.264 6.845v18.388h-9.228V15.075h8.89l.27 4.765c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.751 10.372 6.845 1.685-3.96 5.254-6.846 10.71-6.846 7.343 0 11.249 5.168 11.249 12.483zm34.893 6.308H204.52c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.282 6.601-4.027h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.931 6.938-16.843 16.502-16.843 10.036 0 15.627 6.979 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.436c0-4.16-1.954-6.71-5.86-6.71-4.042 0-6.668 2.751-6.668 6.844v18.388h-9.228V15.075h8.89l.27 4.832c1.953-3.087 5.254-5.167 10.306-5.167 7.476 0 11.518 5.234 11.518 12.616zm15.695-4.966v12.012c0 3.892 1.414 5.838 6.264 5.905h3.619v7.65h-5.034c-9.227-.066-14.078-4.83-14.078-13.22V22.39h-4.849v-7.315h5.388V5.747h8.69v9.328h9.834v7.315h-9.834zM65.02 36.013v-2.08h-4.782c-5.12 0-7.814 1.342-7.814 4.294 0 2.08 1.617 3.423 4.85 3.423 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.785h14.482c-.606-4.697-3.166-6.643-6.938-6.643-3.705 0-6.668 2.08-7.544 6.643zm89.183 0h14.482c-.607-4.697-3.166-6.643-6.938-6.643-3.705 0-6.668 2.08-7.544 6.643zm86.393 12.05h7.66v7.632h-7.66v-7.631z'
-            fill='#FFF'
-            stroke='#FFF'
-            strokeWidth='.6'
-          />
+          <use fill="url(#g)" xlinkHref="#h" />
+          <use fill="url(#i)" xlinkHref="#h" />
         </g>
-      </svg>
-    )
-  }
+        <g transform="translate(4.037 7.058)">
+          <use fill="url(#g)" xlinkHref="#j" />
+          <use fill="url(#k)" xlinkHref="#j" />
+        </g>
+        <g transform="translate(7.062 5.066)">
+          <use fill="url(#g)" xlinkHref="#l" />
+          <use fill="url(#m)" xlinkHref="#l" />
+        </g>
+        <path
+          d="M24.45 31.07c0-5.435-2.829-8.924-7.342-8.924-5.119 0-7.88 4.294-7.88 7.985v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.623 7.41-9.193m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.97-5.167l-.201 4.831H0V0h9.228v19.394c1.953-2.885 5.32-4.966 10.238-4.966 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.483v19.461h-8.554l-.202-4.496c-1.28 2.08-4.378 4.832-10.777 4.832-7.14 0-10.98-3.557-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.692-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.315 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.757zm16.706 19.797c-9.296 0-14.01-3.758-14.078-10.603l7.88.067c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.793-1.14-18.793-14.427 0-6.04 5.39-9.73 14.078-9.73 8.42 0 13.135 3.891 13.202 10.4h-8.016c-.202-2.683-2.02-4.026-5.253-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.739 18.793 14.428.067 6.107-5.524 9.664-14.347 9.664zm47.89-14.764H106.83c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.281 6.6-4.026h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.503-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V28.924c0-4.161-2.087-6.51-5.456-6.51-3.704 0-6.263 2.751-6.263 6.845v18.387h-9.228V28.924c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.751-6.264 6.845v18.387h-9.228V14.764h8.89l.27 4.764c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.752 10.372 6.845 1.685-3.96 5.254-6.845 10.71-6.845 7.342 0 11.249 5.167 11.249 12.482zm34.893 6.308h-23.238c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.281 6.601-4.026h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.037 0-16.368-6.912-16.368-16.71 0-9.932 6.938-16.844 16.502-16.844 10.036 0 15.627 6.98 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.125c0-4.161-1.954-6.711-5.86-6.711-4.042 0-6.669 2.751-6.669 6.845v18.387h-9.227V14.764h8.89l.27 4.832c1.953-3.087 5.254-5.168 10.306-5.168 7.476 0 11.518 5.235 11.518 12.616zm15.694-4.966v12.013c0 3.892 1.415 5.838 6.265 5.905h3.619v7.65h-5.034c-9.228-.067-14.078-4.831-14.078-13.22V22.078h-4.849v-7.314h5.388V5.436h8.69v9.328h9.834v7.314h-9.835zM56.513 35.701v-2.08H51.73c-5.119 0-7.813 1.342-7.813 4.295 0 2.08 1.617 3.422 4.85 3.422 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.784h14.482c-.606-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm89.183 0h14.482c-.607-4.698-3.166-6.644-6.938-6.644-3.705 0-6.668 2.08-7.544 6.644zm86.393 12.05h7.66V47.6h-7.66v-7.632z"
+          transform="translate(5.422 4.311)"
+          fill="url(#n)"
+        />
+        <path
+          d="M32.958 31.382c0-5.435-2.829-8.925-7.342-8.925-5.119 0-7.88 4.295-7.88 7.986v2.08c0 3.96 2.896 8.053 7.813 8.053 4.378 0 7.41-3.624 7.41-9.194m9.362-.403c0 10.2-5.726 17.314-14.684 17.314-4.984 0-8.15-2.08-9.969-5.167l-.202 4.832H8.508V.311h9.228v19.394c1.953-2.885 5.32-4.965 10.238-4.965 8.352 0 14.347 5.838 14.347 16.24zm31.658-2.482v19.46h-8.554l-.202-4.495c-1.28 2.08-4.378 4.831-10.777 4.831-7.14 0-10.98-3.556-10.98-9.06 0-7.985 8.218-10.334 17.177-10.334h4.378v-1.14c0-3.691-2.088-6.04-6.197-6.04-3.503 0-5.793 1.677-6.13 4.563h-8.352c.607-7.314 6.332-11.543 14.752-11.543 9.9 0 14.885 5.302 14.885 13.758zm16.706 19.796c-9.296 0-14.01-3.758-14.078-10.603l7.88.068c.203 2.684 2.022 4.16 6.063 4.16 3.906 0 5.59-1.14 5.523-3.02 0-6.173-18.792-1.14-18.792-14.428 0-6.04 5.388-9.73 14.077-9.73 8.42 0 13.135 3.892 13.202 10.401h-8.015c-.202-2.684-2.021-4.026-5.254-4.026-3.503 0-5.052 1.208-5.052 3.087 0 6.04 18.793.738 18.793 14.428.067 6.107-5.524 9.663-14.347 9.663zm47.89-14.763h-23.237c.673 5.033 3.637 7.583 7.948 7.583 4.445 0 6.062-2.282 6.6-4.027h8.555c-1.078 5.704-5.658 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.931 6.938-16.843 16.503-16.843 10.036 0 15.627 6.979 15.627 17.045v1.745zm54.29-6.308v20.736h-9.295V29.235c0-4.161-2.087-6.51-5.455-6.51-3.705 0-6.264 2.752-6.264 6.845v18.388h-9.228V29.235c0-4.094-2.02-6.51-5.456-6.51-3.705 0-6.264 2.752-6.264 6.845v18.388h-9.228V15.075h8.89l.27 4.765c1.886-3.02 5.12-5.1 9.7-5.1 5.32 0 8.823 2.751 10.372 6.845 1.685-3.96 5.254-6.846 10.71-6.846 7.343 0 11.249 5.168 11.249 12.483zm34.893 6.308H204.52c.674 5.033 3.637 7.583 7.948 7.583 4.446 0 6.062-2.282 6.601-4.027h8.555c-1.078 5.704-5.659 11.207-15.627 11.207-10.036 0-16.368-6.912-16.368-16.71 0-9.931 6.938-16.843 16.502-16.843 10.036 0 15.627 6.979 15.627 17.045v1.745zm34.084-6.174v20.602h-9.228V29.436c0-4.16-1.954-6.71-5.86-6.71-4.042 0-6.668 2.751-6.668 6.844v18.388h-9.228V15.075h8.89l.27 4.832c1.953-3.087 5.254-5.167 10.306-5.167 7.476 0 11.518 5.234 11.518 12.616zm15.695-4.966v12.012c0 3.892 1.414 5.838 6.264 5.905h3.619v7.65h-5.034c-9.227-.066-14.078-4.83-14.078-13.22V22.39h-4.849v-7.315h5.388V5.747h8.69v9.328h9.834v7.315h-9.834zM65.02 36.013v-2.08h-4.782c-5.12 0-7.814 1.342-7.814 4.294 0 2.08 1.617 3.423 4.85 3.423 4.85 0 7.746-3.02 7.746-5.637zm50.451-7.785h14.482c-.606-4.697-3.166-6.643-6.938-6.643-3.705 0-6.668 2.08-7.544 6.643zm89.183 0h14.482c-.607-4.697-3.166-6.643-6.938-6.643-3.705 0-6.668 2.08-7.544 6.643zm86.393 12.05h7.66v7.632h-7.66v-7.631z"
+          fill="#FFF"
+          stroke="#FFF"
+          strokeWidth=".6"
+        />
+      </g>
+    </svg>
+  )
 }
 
-export default LogoMobile
+export default LogoGlitched
