@@ -1,10 +1,10 @@
-import React from 'react'
-import Logo from './Logo'
-import LogoMobile from './LogoMobile'
-import Link from './Link'
+import React from "react"
+import Logo from "./Logo"
+import LogoMobile from "./LogoMobile"
+import Link from "./Link"
 
-class Nav extends React.Component {
-  constructor (props) {
+class Nav extends React.PureComponent {
+  constructor(props) {
     super(props)
     this.state = {
       vpWidth: 1000
@@ -13,64 +13,64 @@ class Nav extends React.Component {
     this.updateWindowDimensions = this.updateWindowDimensions.bind(this)
   }
 
-  componentDidMount () {
+  componentDidMount() {
     this.updateWindowDimensions()
-    window.addEventListener('resize', this.updateWindowDimensions)
+    window.addEventListener("resize", this.updateWindowDimensions)
   }
 
-  updateWindowDimensions () {
+  updateWindowDimensions() {
     this.setState({ vpWidth: window.innerWidth })
   }
 
-  componentWillUnmount () {
+  componentWillUnmount() {
     this.state.vpWidth
-    window.removeEventListener('resize', this.updateWindowDimensions)
+    window.removeEventListener("resize", this.updateWindowDimensions)
   }
 
-  render () {
+  render() {
     const logoAnimation = this.props.animate
       ? {
-        'data-aos': 'fade-down',
-        'data-aos-offset': '50',
-        'data-aos-delay': '200',
-        'data-aos-duration': '1000',
-        'data-aos-easing': 'cubic-bezier(.05, .69, .14, 1)'
-      }
+          "data-aos": "fade-down",
+          "data-aos-offset": "-150",
+          "data-aos-delay": "200",
+          "data-aos-duration": "1000",
+          "data-aos-easing": "cubic-bezier(.05, .69, .14, 1)"
+        }
       : {}
 
     const navAnimation = this.props.animate
       ? {
-        'data-aos': 'fade-down',
-        'data-aos-offset': '50',
-        'data-aos-delay': '500',
-        'data-aos-duration': '1000',
-        'data-aos-easing': 'cubic-bezier(.05, .69, .14, 1)'
-      }
+          "data-aos": "fade-down",
+          "data-aos-offset": "-150",
+          "data-aos-delay": "500",
+          "data-aos-duration": "1000",
+          "data-aos-easing": "cubic-bezier(.05, .69, .14, 1)"
+        }
       : {}
 
     return (
       <nav>
-        <figure className='home' {...logoAnimation}>
-          <Link href='/'>
-            <a className='logo' aria-label='basement'>
+        <figure className="home" {...logoAnimation}>
+          <Link href="/">
+            <a className="logo" aria-label="basement">
               {this.state.vpWidth < 800 ? <LogoMobile /> : <Logo />}
             </a>
           </Link>
         </figure>
         <ul {...navAnimation}>
           <li>
-            <Link activeClassName='active' href='/about' prefetch>
-              <a className='navigation-item'>About</a>
+            <Link activeClassName="active" href="/about" prefetch>
+              <a className="navigation-item">About</a>
             </Link>
           </li>
           <li>
-            <Link activeClassName='active' href='/work' prefetch>
-              <a className='navigation-item'>Work</a>
+            <Link activeClassName="active" href="/work" prefetch>
+              <a className="navigation-item">Work</a>
             </Link>
           </li>
           <li>
-            <Link activeClassName='active' href='/contact' prefetch>
-              <a className='navigation-item'>Contact</a>
+            <Link activeClassName="active" href="/contact" prefetch>
+              <a className="navigation-item">Contact</a>
             </Link>
           </li>
         </ul>
@@ -87,6 +87,13 @@ class Nav extends React.Component {
             display: flex;
             justify-content: space-between;
             align-items: center;
+            width: 100vw;
+            padding: 1.5rem 7.17vw;
+            position: fixed;
+            left: 50%;
+            transform: translate(-50%);
+            background-color: #000;
+            z-index: 1000;
           }
 
           .home {
