@@ -1,32 +1,32 @@
 // next.config.js
-const fs = require('fs')
-const { join } = require('path')
-const { promisify } = require('util')
+const fs = require("fs")
+const { join } = require("path")
+const { promisify } = require("util")
 const copyFile = promisify(fs.copyFile)
-const withOffline = require('next-offline')
+const withOffline = require("next-offline")
 
 const nextConfig = {
-  target: 'serverless',
+  target: "serverless",
   workboxOpts: {
-    swDest: 'static/service-worker.js',
+    swDest: "static/service-worker.js",
     runtimeCaching: [
       {
         urlPattern: /.png$/,
-        handler: 'cacheFirst'
+        handler: "cacheFirst"
       },
       {
         urlPattern: /.jpg$/,
-        handler: 'cacheFirst'
+        handler: "cacheFirst"
       },
       {
         urlPattern: /.svg$/,
-        handler: 'cacheFirst'
+        handler: "cacheFirst"
       },
       {
         urlPattern: /^https?.*/,
-        handler: 'networkFirst',
+        handler: "networkFirst",
         options: {
-          cacheName: 'https-calls',
+          cacheName: "https-calls",
           networkTimeoutSeconds: 15,
           expiration: {
             maxEntries: 150,
@@ -39,13 +39,13 @@ const nextConfig = {
       }
     ]
   },
-  exportPathMap: async function (defaultPathMap, { dev, dir, outDir }) {
+  exportPathMap: async function(defaultPathMap, { dev, dir, outDir }) {
     if (dev) {
       return defaultPathMap
     }
     // This will copy robots.txt from your project root into the out directory
-    await copyFile(join(dir, 'robots.txt'), join(outDir, 'robots.txt'))
-    await copyFile(join(dir, 'sitemap.xml'), join(outDir, 'sitemap.xml'))
+    await copyFile(join(dir, "robots.txt"), join(outDir, "robots.txt"))
+    await copyFile(join(dir, "sitemap.xml"), join(outDir, "sitemap.xml"))
     return defaultPathMap
   }
 }
